+++ include/builtins.h
@@ -1,0 +1,8 @@
+#ifndef BUILTINS_H
+#define BUILTINS_H
+
+int cpsh_history(char **args);
+int cpsh_cd(char **args);
+int cpsh_exit(char **args);
+
+#endif
