+++ include/cpsh.h
@@ -1,0 +1,6 @@
+#ifndef CPSH_H
+#define CPSH_H
+
+void cpsh_loop(void);
+
+#endif
