+++ include/executor.h
@@ -1,0 +1,6 @@
+#ifndef EXECUTOR_H
+#define EXECUTOR_H
+
+int cpsh_execute(char **args);
+
+#endif
