+++ include/prompt.h
@@ -1,0 +1,7 @@
+#ifndef PROMPT_H
+#define PROMPT_H
+
+char *get_cwd(void);
+void cpsh_print_prompt(void);
+
+#endif
