+++ src/main.c
@@ -1,4 +1,4 @@
-#include "../include/shell.h"
+#include "../include/cpsh.h"
 
 int main(int argc, char **argv) {
   cpsh_loop();  // Call the cpsh_loop function
